Leap year check
Python Program to Check Leap Year.

Additional Info about Leap Year:
A year is called a leap year if it contains an additional day which makes the number of the days in that year is 366. This additional day is added in February which makes it 29 days long. A leap year occurred once every 4 years.