Python image classification
Image classification is a method to classify the images into their respective category classes using some method like :

Training a small network from scratch
Fine tuning the top layers of the model using VGG16
Let’s discuss how to train model from scratch and classify the data containing cars and planes.

Train Data : Train data contains the 200 images of each cars and planes i.e. total their are 400 images in the training dataset
Test Data : Test data contains 50 images of each cars and planes i.e. total their are 100 images in the test dataset

Conv2D is the layer to convolve the image into multiple images
Activation is the activation function.
MaxPooling2D is used to max pool the value from the given size matrix and same is used for the next 2 layers. then, Flatten is used to flatten the dimensions of the image obtained after convolving it.
Dense is used to make this a fully connected model and is the hidden layer.
Dropout is used to avoid overfitting on the dataset.
Dense is the output layer contains only one neuron which decide to which category image belongs.