Create junk_file_organizer
Aim
The aim of this project is to build a python program that runs as a command-line tool. Basically, as a lazy programmer, my desktop is full of files as shown below. Due to a large number of files, it is a daunting task to sit and organize each file. To make that task easy, we need a Python script that comes in handy and returns a folder where all the files are organized in a well-manner within seconds. It should organize by size, last used or even extensions.